Implement comprehensive cooldown system with visual feedback
- Support fractional hours for minute-level cooldown granularity
- Show cooldown state immediately after roll confirmation
- Display remaining time directly in disabled button for all devices
- Persist cooldown state across page reloads by calculating from last roll
- Improve mobile UX with embedded time display instead of hover tooltips

+++ backend/app/models.py
@@ -31,7 +31,7 @@ class Decision(SQLModel, table=True):
     user_id: int = Field(foreign_key="user.id")
     title: str = Field(max_length=200)
     type: DecisionType
-    cooldown_hours: int = Field(default=0, ge=0)  # 0 means no cooldown
+    cooldown_hours: float = Field(default=0, ge=0)  # 0 means no cooldown
     created_at: datetime = Field(
         default_factory=lambda: datetime.now(timezone.utc), sa_column=Column(DateTime(timezone=True), nullable=False)
     )
